+++ src/game/physics/Point.java
@@ -2,15 +2,15 @@ package game.physics;
 
 /** circle object used for physics */
 public class Point {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
     
     /**
      * Creates a new Point object at x,y that can be used in physics 
      * @param x
      * @param y
      */
-	public Point(int x,int y) {
+	public Point(double x,double y) {
 		this.x = x;
 		this.y = y;
 	}
